\name{lero.lero}
\alias{lero.lero}
\title{Generate 'Lero Lero' quotes}
\usage{
lero.lero(modo = "original")
}
\arguments{
  \item{modo}{Indicates a mode. One of "original", "psicanalitico", or "programador".}
}
\description{
Generates a quote from 'Lero Lero' database (see <http://www.lerolero.com/> for more information).
}
\examples{
# lero.lero()
# lero.lero(modo = "programador")
# lero.lero(modo = "psicanalitico")
}
